# frontend/src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Settly Core API",
    "version": "1.0.0",
    "description": "Authoritative HTTP API contract for the Settly real estate platform. Implements RFC 9457 error taxonomy, strict validation, and bilingual domain entities."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "settly_session",
        "description": "Better Auth HTTP-only secure session cookie"
      }
    },
    "schemas": {
      "BaseProblem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Relative URI identifier of the specific problem type (machine-readable code)",
            "example": "/errors/validation-failed"
          },
          "title": {
            "type": "string",
            "description": "Short human-readable developer summary of the problem type",
            "example": "Validation Failed"
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code generated by the origin server",
            "example": 422
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence of the problem",
            "example": "One or more request parameters failed validation schema checks."
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem",
            "example": "/api/v1/areas?level=INVALID"
          },
          "requestId": {
            "type": "string",
            "description": "Authoritative request correlation ID (UUIDv7) for telemetry tracing",
            "example": "0191eb45-8f67-73d8-9db8-bc234a9e51c8"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "Structured machine-readable parameters for localized client error rendering"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "requestId"
        ]
      },
      "ValidationProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/validation-failed"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  422
                ]
              },
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string",
                      "example": "level"
                    },
                    "code": {
                      "type": "string",
                      "example": "invalid_enum_value"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid enum value"
                    },
                    "params": {
                      "type": "object",
                      "additionalProperties": {
                        "nullable": true
                      }
                    }
                  },
                  "required": [
                    "path",
                    "code"
                  ]
                },
                "description": "Array of validation issues per field"
              }
            },
            "required": [
              "errors"
            ]
          }
        ]
      },
...
      "ForbiddenProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/forbidden"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  403
                ]
              }
            }
          }
        ]
      },
      "UnauthenticatedProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/unauthenticated"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  401
                ]
              }
            }
          }
        ]
      },
...
      "AmenityItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "0191eb45-8f67-73d8-9db8-bc234a9e51c8"
          },
          "slug": {
            "type": "string",
            "example": "private-pool"
          },
          "nameEn": {
            "type": "string",
            "example": "Private Pool"
          },
          "nameAr": {
            "type": "string",
            "example": "حمام سباحة خاص"
          },
          "icon": {
            "type": "string",
            "nullable": true,
            "example": "pool"
          },
          "category": {
            "type": "string",
            "enum": [
              "INTERIOR",
              "EXTERIOR",
              "FACILITY",
              "SECURITY",
              "LOCATION"
            ],
            "example": "EXTERIOR"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-05T12:00:00Z"
          }
        },
        "required": [
          "id",
          "slug",
          "nameEn",
          "nameAr",
          "icon",
          "category",
          "createdAt"
        ]
      },
...
      "AmenityListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AmenityItem"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CreateAmenity": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "nameEn": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "nameAr": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "icon": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "INTERIOR",
              "EXTERIOR",
              "FACILITY",
              "SECURITY",
              "LOCATION"
            ],
            "default": "INTERIOR"
          }
        },
        "required": [
          "slug",
          "nameEn",
          "nameAr"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/amenities": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List property amenities",
        "description": "Retrieve controlled vocabulary of property amenities, optionally filtered by category.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "INTERIOR",
                "EXTERIOR",
                "FACILITY",
                "SECURITY",
                "LOCATION"
              ]
            },
            "required": false,
            "name": "category",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of amenities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmenityListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog",
          "Admin"
        ],
        "summary": "Create property amenity (Admin only)",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAmenity"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Amenity created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmenityItem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenProblem"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
...
  }
}